--- src/DefectLoginPage/Custominput.js
@@ -1,38 +1,0 @@
-import { useField , useFormikContext } from "formik"
-import { TextField,Box, FormControl } from "@mui/material"
-import { useEffect } from "react"
-
-const OutlineColor =(color) => {
-    return {"& .MuiOutlinedInput-root .MuiOutlinedInput-notchedOutline": { borderColor: color }}}       
-
-
-const CustomInput = ({label,extraOnChange,style,...props}) => {
-
-    const[field,meta]=useField(props)
-    const { setFieldValue } = useFormikContext();
-
-    const newOnChange=(event)=>{
-        setFieldValue(props.name,props.value)
-        extraOnChange(event)
-    }
-
-    useEffect(() => {
-        setFieldValue(props.name,props.value)
-    },[props.value])
-    
-    return(
-    <> 
-            <TextField 
-            sx={{...style ,...(meta.touched && meta.error ? OutlineColor("#ff0000") : null) 
-            }}
-            color="third"
-            size="small" 
-            {...field} {...props}
-            onChange={newOnChange}
-            />
-    
-            
-    </>
-    )
-}
-export default CustomInput
